--- a/schema.sql
+++ b/schema.sql
@@ -56,3 +56,11 @@ CREATE TABLE visits (
 ALTER TABLE owners ADD COLUMN email VARCHAR(120);
 
 ALTER TABLE visits ALTER COLUMN date_of_visit TYPE timestamp;
+
+CREATE INDEX animal_id_asc ON visits(animal_id ASC);
+
+CREATE INDEX vet_id_asc ON visits(vet_id ASC);
+
+DROP INDEX vet_id_asc;
+
+CREATE INDEX email_asc ON owners(email ASC);