--- a/schema.sql
+++ b/schema.sql
@@ -10,7 +10,6 @@ CREATE TABLE animals (
   neutered boolean,
   weight_kg decimal
 );
-
 ALTER TABLE animals ADD species varchar(100);
 
 CREATE TABLE owners (