- EXPLAIN ANALYZE used on queries to check the performance;
- Non-clustered indexes were created to decrease the execution time;
- All changes on the database schema were added to the schema.sql file;
- The INSERT commands used were added to the data.sql file;

### Results


**`SELECT COUNT(*) FROM visits where animal_id = 4;`**

- Before improving the speed

<img width="1002" alt="Screen Shot 2022-08-15 at 4 34 52 PM" src="https://user-images.githubusercontent.com/47673311/184725272-de266705-aeac-429d-878a-d8a00ebcad8f.png">
<img width="1002" alt="Screen Shot 2022-08-15 at 4 35 21 PM" src="https://user-images.githubusercontent.com/47673311/184725415-286dbf86-dd07-444d-8b2a-51df652b6766.png">

- After creating a non-clustered index for `animal_id` on the table `visits`

<img width="1185" alt="Screen Shot 2022-08-15 at 4 37 33 PM" src="https://user-images.githubusercontent.com/47673311/184725508-a974b7b4-47f9-46eb-8b0a-3e7d427124e0.png">
<img width="1185" alt="Screen Shot 2022-08-15 at 4 39 45 PM" src="https://user-images.githubusercontent.com/47673311/184725550-ee09fff2-ad38-454a-bb95-deb2d977a987.png">

- Conclusion: The average execution time, of the three measures, decreased from 1,241.9 ms to 153.7 ms. The creation of the index for animal_id had a great impact on improving performance.

\
**`SELECT * FROM visits where vet_id = 2;`**

- Before improving the speed

<img width="1060" alt="Screen Shot 2022-08-15 at 12 40 11 PM" src="https://user-images.githubusercontent.com/47673311/184726775-0a2ba9a6-37a9-42f9-85aa-bdbe73f3956d.png">
<img width="1016" alt="Screen Shot 2022-08-15 at 4 44 02 PM" src="https://user-images.githubusercontent.com/47673311/184726990-43188134-0530-4911-9f3c-a7337f0121d5.png">

- After creating a non-clustered index for `vet_id` on the table `visits`

<img width="1016" alt="Screen Shot 2022-08-15 at 6 33 44 PM" src="https://user-images.githubusercontent.com/47673311/184727101-8c8c5833-7ecb-4d07-9c24-7618bf15125a.png">

- Conclusion: The average execution time, of the three measures, decreased from 687.4 ms to 367.8 ms. Creating the index for vet_id had no performance impact. The vet_id is already in the primary key as the first column. Only the first query has a high execution time due to caching.

\
**`SELECT * FROM owners where email = 'owner_18327@mail.com';`**

- Before improving the speed

<img width="1016" alt="Screen Shot 2022-08-15 at 4 45 22 PM" src="https://user-images.githubusercontent.com/47673311/184729581-ffa21d5d-d018-4b83-9228-10a7671db953.png">

- After creating a non-clustered index for `email` on the table `owners`

<img width="1016" alt="Screen Shot 2022-08-15 at 6 35 38 PM" src="https://user-images.githubusercontent.com/47673311/184730418-34248fb7-8044-4663-a303-8bde48178d9d.png">

- Conclusion: The average execution time, of the three measures, decreased from 1,700.6 ms to 0.6 ms. The creation of the index for email had a great impact on improving performance.